backend/api/drf/requirements.txt
asgiref==3.2.3
autopep8==1.5
Django==3.0.3
djangorestframework==3.11.0
pkg-resources==0.0.0
pycodestyle==2.5.0
Pygments==2.5.2
pylint-django==2.0.14
pylint-plugin-utils==0.6
pytz==2019.3
sqlparse==0.3.0
